Move interface IObjectPrx and class ObjectPrx into IObjectPrx.cs
